+++ core/src/com/color/game/enums/UserDataType.java
@@ -3,6 +3,7 @@ package com.color.game.enums;
 
 public enum UserDataType {
 
+    TELEPORTER,
     EXIT,
     NOTICE,
     PLATFORM,